Changes:  - data library component name() member functions now return    core::identifier_string instead of std::string  - fix for compiling specification_property_map with gcc 4.1.2 on Linux
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6315 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/lps/include/mcrl2/lps/detail/specification_property_map.h
@@ -334,7 +334,7 @@ namespace detail {
         unsigned int result = 0;
         for (std::map<std::string, std::string>::const_iterator i = m_data.begin(); i != m_data.end(); ++i)
         {
-          result = (std::max)(result, i->first.size());
+          result = (std::max)(static_cast< size_t >(result), i->first.size());
         }
         return result;
       }
